% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{captcha_accuracy_oracle}
\alias{captcha_accuracy_oracle}
\title{Captcha accuracy metric with oracle}
\usage{
captcha_accuracy_oracle()
}
\description{
This object is used to calculate the accuracy of the model in the
fitting process.
}
\details{
This function is a generator created using \code{\link[luz:luz_metric]{luz::luz_metric()}} function.
It has a \code{initialize()} method that sets the total number of instances
and total number of correct predictions as zero. For any minibatch, it
has an \code{update()} method that updates the total number of instances and
total number of correct predictions with new data. Finally, it has a
\code{compute()} method that calculates accuracy from the total number of
correct predictions and total number of instances.

Training data has a different structure than validation
data: training data includes both complete and incomplete labels
from the oracle, but validation includes only complete labels. Accuracy
is calculated using only when we have at least one complete label in the
minibatch. Accuracy is calculated considering only the complete labels.
}
