% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle.R
\name{captcha_candidates}
\alias{captcha_candidates}
\title{Calculates n captcha candidate labels}
\usage{
captcha_candidates(f_captcha, model, cut_value = log(0.01), n)
}
\arguments{
\item{f_captcha}{captcha file}

\item{model}{model that generates logits}

\item{cut_value}{so that we make combinations from real values}

\item{n}{maximum number of candidates}
}
\description{
This function returns a list of candidated using all captcha combinations
from some cut value. It uses probabilities returned from the model to get
the most likely labels.
}
