% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{captcha_dataset_oracle}
\alias{captcha_dataset_oracle}
\title{Captcha dataset incorporating incomplete data structure}
\usage{
captcha_dataset_oracle(
  root,
  path_logs,
  transform_image = captcha::captcha_transform_image,
  transform_label = captcha::captcha_transform_label,
  augmentation = NULL
)
}
\arguments{
\item{root}{(string): root directory of the files}

\item{path_logs}{path to log files}

\item{transform_image}{(callable, optional): A function/transform
that takes in an file path and returns an torch tensor prepared
to feed the model.}

\item{transform_label}{(callable, optional): A function/transform
that takes in the file paths and transform them.}

\item{augmentation}{(function, optional) If not \code{NULL}, applies a
function to augment data with randomized preprocessing layers.

This is an object of class \code{dataset_generator} created using
\code{\link[torch:dataset]{torch::dataset()}} function. It has a \code{initialize()} method that
takes directory containing the input images and the log file,
then assigns all the information in-memory with the list-like data
structure for the response variable. It also assigns a dummy variable
that indicates whether the information is incomplete, to facilitate
loss and accuracy calculation. It also has a \code{.getitem()} method that
correctly extracts one observation of the dataset in this complex
structure, and a \code{.length()} method that correctly calculates the
number of Captchas of the dataset.}
}
\description{
When we use oracles, we have to deal with a different data structure
that incorporates incomplete information provided by the websites.
First, we need to pass not only the directory of the images but also
a \code{log} file that records all the attempts to solve the Captcha through
the website. When there is at least one successful attempt, the image
and the label are read normally, as if we were in the complete data
framework. However, when all the attempts fail, we need to record all
the failed attempts (a list of one-hot encoded labels) for this observation.
The number of failed attempts is not fixed, then we have to store the
data as a list.
}
