% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{captcha_dataset_oracle_online}
\alias{captcha_dataset_oracle_online}
\title{Captcha dataset incorporating online learning}
\usage{
captcha_dataset_oracle_online(
  root,
  model,
  ntry = 1,
  captcha_access_fn,
  captcha_test_fn,
  p_new = 0.2
)
}
\arguments{
\item{root}{(string): root directory to save new files}

\item{model}{initial model.}

\item{ntry}{number of attempts to solve the Captcha. Defaults to 1.}

\item{captcha_access_fn}{function that downloads a Captcha image and returns
all the information needed to test whether the label is correct.}

\item{captcha_test_fn}{function that uploads a prediction and tests whether it
is correct, returning \code{TRUE} or \code{FALSE}.}

\item{p_new}{hyperparameter to control the probability of downloading new
data to get a new observation. For example, when \code{p_new} is \code{0.2}, the
default, approximately 20\% of the times we get an image to compose a
minibatch it is a new image downloaded from the internet, and 80\% of the
times it is an already downloaded image.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The difference from this function to \code{\link[=captcha_dataset_oracle]{captcha_dataset_oracle()}} is that
it downloads the new data directly from the internet instead of considering
a fixed dataset. To take advantage of the fact that we are downloading a lot
of new images (and new information), it is possible to consider a
probability to download a new image every time we need to get an
observation. This way the downloaded images can be used in more than one
minibatch and we will always have the chance to get new information.
}
\details{
This \code{dataset_generator} object implements an experimental feature to
learn how to solve the Captcha automatically accessing the web using
an oracle function. It does not need any input data, only the initial
model of class \code{luz_module_fitted} and the \code{captcha_access_fn} and
\code{captcha_test_fn} functions to access the web.
}
