% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{captcha_dataset_oraculo}
\alias{captcha_dataset_oraculo}
\title{Captcha dataset do oráculo}
\usage{
captcha_dataset_oraculo(
  root,
  path_logs,
  transform_image = captcha::captcha_transform_image,
  transform_label = captcha::captcha_transform_label,
  augmentation = NULL
)
}
\arguments{
\item{root}{(string): root directory of the files}

\item{path_logs}{path to log files}

\item{transform_image}{(callable, optional): A function/transform
that takes in an file path and returns an torch tensor prepared
to feed the model.}

\item{transform_label}{(callable, optional): A function/transform
that takes in the file paths and transform them.}

\item{augmentation}{(function, optional) If not \code{NULL}, applies a
function to augment data with randomized preprocessing layers.}
}
\description{
Captcha dataset do oráculo
}
