% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle.R
\name{captcha_oracle}
\alias{captcha_oracle}
\title{Function to download a captcha and classify using the Oracle}
\usage{
captcha_oracle(
  path,
  model = NULL,
  max_ntry = 10,
  manual = TRUE,
  captcha_access,
  captcha_test
)
}
\arguments{
\item{path}{path of the directory to save files}

\item{model}{model used to predict the label of an image. Defaults to NULL,
which opens a prompt to manual input.}

\item{max_ntry}{maximum number of attempts. Defaults to 10.}

\item{manual}{if all the attempts fail, should it opem the prompt for
manual input? Defaults to \code{TRUE}.}

\item{captcha_access}{function that downloads a Captcha image and returns
all the information needed to test whether the label is correct.}

\item{captcha_test}{function that uploads a prediction and tests whether it
is correct, returning \code{TRUE} or \code{FALSE}.}
}
\value{
path to the log file.
}
\description{
This function creates annotated datasets from an initial
model and functions to access the oracle of a website. It stores the
downloaded image and a \code{log} file containing all the attempts to
solve the Captcha.
}
