% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captchaOracle-package.R
\docType{package}
\name{captchaOracle-package}
\alias{captchaOracle}
\alias{captchaOracle-package}
\title{captchaOracle: Modelo Usando Oraculo}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jtrecenti/captchaOracle}
  \item Report bugs at \url{https://github.com/jtrecenti/captchaOracle/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
