% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{collate_oracle}
\alias{collate_oracle}
\title{Collect and bind minibatches with oracle structure}
\usage{
collate_oracle(mb)
}
\arguments{
\item{mb}{the minibatch, which is a list of torch tensors.}
}
\description{
This function reimplements the default \code{collate_fn} function from
torch dataloaders to deal with the list-like structure of datasets
generated from the \code{\link[=captcha_dataset]{captcha_dataset()}} module. Whenever possible,
it applies the \code{\link[torch:torch_stack]{torch::torch_stack()}} function to bind the tensors.
}
