% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{collate_oraculo}
\alias{collate_oraculo}
\title{Função para juntar os dados de um minibatch corretamente}
\usage{
collate_oraculo(l)
}
\arguments{
\item{l}{lista proveniente de um minibatch}
}
\description{
Função para juntar os dados de um minibatch corretamente
}
