% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{net_captcha_oracle}
\alias{net_captcha_oracle}
\title{Captcha model using initial model}
\usage{
net_captcha_oracle(
  base_model = NULL,
  input_dim,
  output_ndigits,
  output_vocab_size,
  vocab,
  transform,
  dropout = c(0.25, 0.25),
  dense_units = 400
)
}
\arguments{
\item{base_model}{base model of class \code{luz_module_fitted}. Defaults to
\code{NULL}, which indicates no initialization of the model parameters.}

\item{input_dim}{input image dimensions}

\item{output_ndigits}{response variable number of digits}

\item{output_vocab_size}{integer indicating size of the vocabulary}

\item{vocab}{vocabulary elements}

\item{transform}{function to transform input data}

\item{dropout}{vector of two elements indicating dropout hyperparameter}

\item{dense_units}{integer indicating number of units used in the
dense layer applied after the convolutional layers}
}
\value{
object of classes \code{ORACLE-CNN} and \code{nn_module}. It works
as a predictive function and as the input to a luz fitting workflow.
}
\description{
This is the same as the original model, but including an initial
model to improve using new data. It only uses the input data, so
we do not need to adapt anything from the original function here.
}
