% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{captcha_dataset_oraculo}
\alias{captcha_dataset_oraculo}
\title{Captcha dataset do oráculo}
\usage{
captcha_dataset_oraculo(
  root,
  captcha,
  path_logs,
  transform_image = captcha::captcha_transform_image,
  transform_label = captcha::captcha_transform_label,
  download = FALSE,
  in_memory = TRUE,
  augmentation = NULL
)
}
\arguments{
\item{root}{(string): root directory of dataset where \code{captcha.zip}
exists or will be saved to if download is set to \code{TRUE}}

\item{captcha}{(string): name of the captcha, must be one of
\url{available_captchas()}}

\item{path_logs}{path to log files}

\item{transform_image}{(callable, optional): A function/transform
that takes in an file path and returns an torch tensor prepared
to feed the model.}

\item{transform_label}{(callable, optional): A function/transform
that takes in the file paths and transform them.}

\item{download}{(bool, optional): If \code{TRUE}, downloads the dataset
from the internet and puts it in \code{root}. If dataset is already
downloaded, it is not downloaded again. Defaults to \code{FALSE}}

\item{in_memory}{(bool, optional) If \code{TRUE}, the default, loads
all the files in memory. If \code{FALSE}, it exports a data generator
function to read batches from disk.}

\item{augmentation}{(function, optional) If not \code{NULL}, applies a
function to augment data with randomized preprocessing layers.}
}
\description{
Captcha dataset do oráculo
}
