% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{net_captcha_oracle}
\alias{net_captcha_oracle}
\title{Modelo de captcha com oraculo}
\usage{
net_captcha_oracle(
  base_model,
  input_dim,
  output_ndigits,
  output_vocab_size,
  vocab,
  transform,
  dropout = c(0.25, 0.25),
  dense_units = 400
)
}
\arguments{
\item{base_model}{modelo base carregado com \code{luz::load_model()}}

\item{input_dim}{dimensão de entrada}

\item{output_ndigits}{quantidade de digitos da resposta}

\item{output_vocab_size}{tamanho do vocabulario da resposta}

\item{vocab}{o vocabulario da resposta}

\item{transform}{função de transformação dos dados}

\item{dropout}{hiperparâmetro de dropout}

\item{dense_units}{hiperparâmetro de dense units na camada densa
que vem depois das convoluções}
}
\description{
É o mesmo modelo que o original, com a diferença que inicializamos
o modelo com os pesos do modelo base ajustado no passo anterior.
}
