% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{oracle_loss}
\alias{oracle_loss}
\title{Loss do oraculo}
\usage{
oracle_loss(weight = NULL, reduction = "mean")
}
\arguments{
\item{weight}{pesos a serem aplicados nas observações}

\item{reduction}{função de redução}
}
\description{
Loss do oraculo
}
